splitting off organism defn
