+++ cleopatra.py
@@ -1,61 +1,10 @@
 #!/usr/bin/python
 
-import random
-
-gene_names = ['A', 'B', 'a', 'b']
-number_of_genes = 100000 
-genotype_dict = {}
-
-def cross(genotype, other_genotype):
-    if genotype > other_genotype:
-      genotype_code = other_genotype + '_' + genotype
-    else:
-      genotype_code = genotype + '_' + other_genotype
-
-    if genotype_code in genotype_dict:
-      new_genotypes = genotype_dict[genotype_code]
-    else:
-      new_genotypes = []
-      for gene_a in [genotype[0],genotype[1]]:
-        for gene_b in [other_genotype[0],other_genotype[1]]:
-          if gene_a > gene_b:
-            new_genotypes.append(gene_b+gene_a)
-          else:
-            new_genotypes.append(gene_a+gene_b)
-
-      genotype_dict[genotype_code] = new_genotypes
-
-    i = random.randint(0,3)
-	 
-    return new_genotypes[i]
-
-class Organism:
-  def __init__(self, name, genelist=[], mother=None, father=None ):
-    self.name = name
-
-    if genelist:
-      self.genes = genelist
-    elif mother and father:
-      self.genes = []
-
-      for i in xrange(number_of_genes):
-        self.genes.append(cross(father.genes[i], mother.genes[i]))
-
-  def show(self):
-    r_dict = {'AA': 0, 'aa': 0, 'BB': 0, 'bb': 0}
-
-    for gene in self.genes:
-       if gene in r_dict:
-         r_dict[gene] += 1
-
-    total = sum(r_dict.values())*100.0/number_of_genes
-
-    print self.name + " was " + str(total) + "% inbred"
-
+from organism import *
          
-Pt5 = Organism("Ptolmey V", genelist=['Aa' for i in xrange(number_of_genes)])
+Pt5 = Organism("Ptolmey V")
 Pt5.show()
-Cl1 = Organism("Cleopatra", genelist=['Bb' for i in xrange(number_of_genes)])
+Cl1 = Organism("Cleopatra")
 Cl1.show()
 
 Pt6 = Organism("Ptolemy VI", mother=Cl1, father=Pt5)
